requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
django==1.11.4
six==1.10.0
djangorestframework==3.3.2
requests==2.14.2
MySQL-python
